requirements.txt
Flask
numpy
mysql-connector-python==8.0.23
Flask-Session==0.3.2
passlib
requests